requirements.txt
pymysql==1.0.2
python-dotenv==1.0.0
paho-mqtt==2.1.0
requests==2.26.0